💄 홈페이지의 제품 카드에 skeleton ui 추가 / 🎨 product data 가져와서 redux store에 저장하는 코드 App.tsx에서 분리
